% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Imputation.R
\name{Impute}
\alias{Impute}
\title{Imputation of missing values of a numerical vector}
\usage{
Impute(x, y, z)
}
\arguments{
\item{x}{is a data frame.}

\item{y}{is the number of the column of the dataframe we want to impute.}

\item{z}{is the desired number to impute with.}
}
\value{
The imputed vector whose missing values are replaced by the desired number
}
\description{
Replace missing values (NA) of a numerical vector with the desired number.
}
\examples{
Impute(data.frame(matrix(c(25,42,NA,12,NA,28,7,18,NA,22,NA,14,14,NA,20),ncol=3)),1,32)
}
